lil chng task11

+++ sem1/pract8/task11.c
@@ -2,18 +2,18 @@
 #define ROWS 3
 #define COLS 5
 
-void modify_arr(int rows, int cols, double arr[rows][cols]) {
-  double *ptr = &arr[0][0];
+void modify_arr(int rows, int cols, int arr[rows][cols]) {
+  int *ptr = &arr[0][0];
   for (int i=0; i<rows*cols; i++, ptr++) *ptr = 2*(*ptr);
 }
 
-void print_arr(int rows, int cols, double arr[rows][cols]) {
-  double *ptr = &arr[0][0];
+void print_arr(int rows, int cols, int arr[rows][cols]) {
+  int *ptr = &arr[0][0];
   printf("{\n");
   for (int i=0; i<rows; i++) {
     printf("  {");
     for (int j=0; j<cols; j++, ptr++) {
-      printf("%.2lf", *ptr);
+      printf("%d", *ptr);
       if (j<cols-1) printf(", ");
     }
     printf("}");
@@ -25,12 +25,12 @@ void print_arr(int rows, int cols, double arr[rows][cols]) {
 int main() {
   printf("The array will be 3x5.\n");
 
-  double arr[ROWS][COLS];
-  double *ptr = &arr[0][0];
+  int arr[ROWS][COLS];
+  int *ptr = &arr[0][0];
 
   for (int i=0; i<ROWS; i++) {
     printf("Enter every element of the %dth array: ", i+1);
-    for (int j=0; j<COLS; j++, ptr++) scanf("%lf", ptr);
+    for (int j=0; j<COLS; j++, ptr++) scanf("%d", ptr);
   }
 
   printf("The array:\n"); print_arr(ROWS, COLS, arr);
